backend/requirements.txt
fastapi
fastapi_jwt_auth
celery
psycopg2-binary
psycopg2-pool
uvicorn
pydantic
python-multipart
shutil